Update tests to match frontend query structure
