--- drivers/hmis/spec/requests/hmis/ce/ce_opportunity_spec.rb
@@ -1,376 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-require_relative '../login_and_permissions'
-require_relative '../../../support/hmis_base_setup'
-
-RSpec.describe Hmis::GraphqlController, type: :request do
-  include_context 'hmis base setup'
-
-  before(:each) do
-    allow_any_instance_of(Hmis::Ce::Configuration).to receive(:enabled?).and_return(true)
-    hmis_login(user)
-  end
-
-  # Basic setup
-  let!(:project) { create :hmis_hud_project, data_source: ds1, user: u1, project_type: 1 }
-  let!(:candidate_pool) { create :hmis_ce_match_candidate_pool }
-  let!(:unit) { create(:hmis_unit_in_group, project: project) }
-  let!(:opportunity) { create :hmis_ce_opportunity, project: project, data_source: ds1, candidate_pool: candidate_pool, unit: unit }
-
-  let!(:access_control) { create_access_control(hmis_user, project, with_permission: [:can_view_project, :can_view_units, :can_view_prioritized_client_lists, :can_view_referrals]) }
-
-  # TODO: adjust this test to test the unit query, since unit will back the "opportunity page"
-  # ```
-  # unit(id: $id) {
-  #    id
-  #    eligibilityRequirements # eligibility requirements independent of opportunity existence
-  #    latestOpportunity {
-  #       referral {}
-  #       candidates {}
-  #    }
-  #    opportunities # history of all opportunities (and referrals) for this unit (maybe later)
-  # }
-  # ```
-  describe 'ce_opportunity query' do
-    let(:query) do
-      <<~GRAPHQL
-        query GetCeOpportunity($id: ID!) {
-          ceOpportunity(id: $id) {
-            id
-            name
-            status
-            expiresAt
-            prioritySchemes { id name expression }
-            candidates {
-              nodesCount
-              nodes {
-                id
-                destinationClientId
-                priorityScores
-                clientName
-              }
-            }
-            referral {
-              id
-              status
-            }
-            eligibilityRequirements {
-              id
-              name
-              ownerType
-              expression
-              projectTypes
-              funders
-            }
-            candidatesGeneratedAt
-          }
-        }
-      GRAPHQL
-    end
-
-    let(:variables) do
-      {
-        id: opportunity.id,
-      }
-    end
-
-    context 'when the opportunity has rules' do
-      let!(:opportunity) { create :hmis_ce_opportunity, project: project, data_source: ds1, candidate_pool: nil, unit: unit }
-      let!(:rule1) { create(:hmis_ce_eligibility_requirement, owner: unit.unit_group) }
-      let!(:rule2) { create(:hmis_ce_eligibility_requirement, owner: project) }
-      let!(:rule3) { create(:hmis_ce_eligibility_requirement, owner: project.organization, applicability_config: { project_types: [project.project_type] }) }
-
-      let!(:funder) { create(:hmis_hud_funder, project: project, data_source: project.data_source) }
-      let!(:rule4) { create(:hmis_ce_eligibility_requirement, owner: project.organization, applicability_config: { project_funders: [funder.funder] }) }
-      before do
-        # Ensure the unit group gets a pool and the opportunity captures assignment rules
-        Hmis::Ce::Match::CandidatePoolBuilder.call
-        # Recreate the opportunity with assignment_rules captured from the resolver
-        # so the GraphQL layer can return historical rules regardless of current state.
-        opportunity.update!(assignment_rules: [rule1, rule2, rule3, rule4].map(&:attributes))
-      end
-
-      it 'returns rules with their correct ownerTypes' do
-        response, result = post_graphql(**variables) { query }
-        expect(response.status).to eq(200), result.inspect
-
-        rules = result.dig('data', 'ceOpportunity', 'eligibilityRequirements')
-        expect(rules).to contain_exactly(
-          a_hash_including('id' => "#{opportunity.id}.#{rule1.id}", 'ownerType' => 'UNIT_GROUP'),
-          a_hash_including('id' => "#{opportunity.id}.#{rule2.id}", 'ownerType' => 'PROJECT'),
-          a_hash_including('id' => "#{opportunity.id}.#{rule3.id}", 'ownerType' => 'ORGANIZATION', 'projectTypes' => ['ES_NBN']),
-          a_hash_including('id' => "#{opportunity.id}.#{rule4.id}", 'ownerType' => 'ORGANIZATION', 'funders' => ['HUD_HUD_VASH']),
-        )
-      end
-
-      it 'returns most-specific priority schemes ordered by rank, then id' do
-        # Build assignment rules with mixed owners and ranks
-        org_rule = create(:hmis_ce_priority_scheme, owner: project.organization, expression: 'org_expr', priority_rank: 2)
-        proj_rule_low = create(:hmis_ce_priority_scheme, owner: project, expression: 'proj_low', priority_rank: 2)
-        proj_rule_high = create(:hmis_ce_priority_scheme, owner: project, expression: 'proj_high', priority_rank: 1)
-        ds_rule = create(:hmis_ce_priority_scheme, owner: project.data_source, expression: 'ds_expr', priority_rank: 1)
-
-        # Keep eligibility rules as before and include priority rules in the snapshot
-        snapshot_rules = [rule1, rule2, rule3, rule4, org_rule, proj_rule_low, proj_rule_high, ds_rule]
-        opportunity.update!(assignment_rules: snapshot_rules.map(&:attributes))
-
-        response, result = post_graphql(**variables) { query }
-        expect(response.status).to eq(200), result.inspect
-
-        prios = result.dig('data', 'ceOpportunity', 'prioritySchemes')
-        # Should only include project-level rules, ordered by priority_rank then id
-        expect(prios.map { |r| r['expression'] }).to eq(['proj_high', 'proj_low'])
-      end
-    end
-
-    context 'when the opportunity has candidates and an active referral' do
-      let!(:ds_access_control) do
-        create_access_control(
-          hmis_user,
-          ds1,
-          with_permission: [
-            :can_view_clients,
-            :can_view_project,
-            :can_view_enrollment_details,
-            :can_view_referrals,
-          ],
-        )
-      end
-
-      let!(:timestamp) { 2.minutes.ago }
-      let(:candidate_pool) { create :hmis_ce_match_candidate_pool, candidates_generated_at: timestamp }
-
-      let!(:client_1) do
-        create(:hmis_hud_client_with_warehouse_client, data_source: ds1)
-      end
-      let!(:client_2) do
-        create(:hmis_hud_client_with_warehouse_client, data_source: ds1)
-      end
-
-      # Create candidates in the pool
-      let!(:candidate1) do
-        create(
-          :hmis_ce_match_candidate,
-          candidate_pool: candidate_pool,
-          client: client_1.destination_client,
-          priority_score: 80,
-        )
-      end
-      let!(:candidate2) do
-        create(
-          :hmis_ce_match_candidate,
-          candidate_pool: candidate_pool,
-          client: client_2.destination_client,
-          priority_score: 100,
-        )
-      end
-
-      let!(:client_with_active_referral) do
-        create(:hmis_hud_client_with_warehouse_client, data_source: ds1)
-      end
-      # Create an active referral for one client
-      let!(:active_referral) do
-        create(
-          :hmis_ce_referral,
-          opportunity: opportunity,
-          data_source: ds1,
-          client: client_with_active_referral,
-          status: 'in_progress',
-        )
-      end
-
-      context 'when querying an opportunity' do
-        it 'returns basic opportunity expected results' do
-          response, result = post_graphql(**variables) { query }
-          expect(response.status).to eq(200), result.inspect
-
-          opportunity_data = result.dig('data', 'ceOpportunity')
-
-          expect(opportunity_data).to include(
-            'id' => opportunity.id.to_s,
-            'name' => opportunity.name,
-            'status' => opportunity.status,
-            'candidatesGeneratedAt' => timestamp.iso8601,
-          )
-
-          candidates = result.dig('data', 'ceOpportunity', 'candidates', 'nodes')
-
-          expect(candidates).to be_an(Array)
-          expect(candidates.size).to eq(2)
-
-          # Verify candidates are ordered by priority score
-          expect(candidates.map { |c| c['priorityScores'] }).to eq([[100], [80]])
-
-          # Verify first candidate (highest priority)
-          expect(candidates[0]).to include(
-            'priorityScores' => [100],
-            'destinationClientId' => client_2.destination_client.id.to_s,
-          )
-
-          # Verify second candidate
-          expect(candidates[1]).to include(
-            'priorityScores' => [80],
-            'destinationClientId' => client_1.destination_client.id.to_s,
-          )
-        end
-
-        it 'returns the active referral' do
-          response, result = post_graphql(**variables) { query }
-          expect(response.status).to eq(200), result.inspect
-
-          referral_data = result.dig('data', 'ceOpportunity', 'referral')
-
-          expect(referral_data).to include(
-            'id' => active_referral.id.to_s,
-            'status' => 'in_progress',
-          )
-        end
-
-        it 'returns no candidates when user lacks permission to see prioritized client lists' do
-          remove_permissions(access_control, :can_view_prioritized_client_lists)
-          response, result = post_graphql(**variables) { query }
-          expect(response.status).to eq(200), result.inspect
-          expect(result.dig('data', 'ceOpportunity', 'candidates', 'nodes')).to be_empty
-        end
-      end
-    end
-
-    describe 'when the opportunity has several referrals' do
-      let!(:workflow_template) { create(:hmis_workflow_definition_template, data_source: ds1) }
-      let!(:opportunity) { create :hmis_ce_opportunity, project: project, data_source: ds1, candidate_pool: candidate_pool, unit: unit, workflow_template: workflow_template }
-      # opportunities are single-use, so there should only be one in-progress or accepted referral, but there could be many failed referrals.
-      let!(:rejected1) { create(:hmis_ce_referral, opportunity: opportunity, data_source: ds1, status: 'rejected', created_at: 1.day.ago) }
-      let!(:rejected2) { create(:hmis_ce_referral, opportunity: opportunity, data_source: ds1, status: 'rejected', created_at: 1.day.ago) }
-      let!(:rejected3) { create(:hmis_ce_referral, opportunity: opportunity, data_source: ds1, status: 'rejected', created_at: 1.day.ago) }
-
-      ['initialized', 'in_progress', 'accepted'].each do |status|
-        it "returns the #{status} referral" do
-          # it should return this referral even if it was created less recently than the rejected referrals (which should not happen)
-          referral = create(:hmis_ce_referral, opportunity: opportunity, data_source: ds1, status: status, created_at: 2.days.ago)
-          response, result = post_graphql(**variables) { query }
-          expect(response.status).to eq(200), result.inspect
-          referral_data = result.dig('data', 'ceOpportunity', 'referral')
-          expect(referral_data).to include('id' => referral.id.to_s, 'status' => status)
-        end
-      end
-    end
-
-    context 'when the opportunity has no candidates' do
-      let(:empty_pool_opportunity) { create :hmis_ce_opportunity, project: project, data_source: ds1 }
-
-      let(:empty_pool_variables) do
-        {
-          id: empty_pool_opportunity.id,
-        }
-      end
-
-      it 'returns an empty candidates array' do
-        response, result = post_graphql(**empty_pool_variables) { query }
-        expect(response.status).to eq(200), result.inspect
-
-        candidates = result.dig('data', 'ceOpportunity', 'candidates', 'nodes')
-
-        expect(candidates).to be_an(Array)
-        expect(candidates).to be_empty
-      end
-    end
-
-    context 'when the opportunity has lots of candidates' do
-      before do
-        50.times do
-          client = create(:hmis_hud_client_with_warehouse_client, data_source: ds1)
-          create(:hmis_ce_match_candidate, candidate_pool: candidate_pool, client: client.destination_client, priority_score: rand(80..100))
-        end
-      end
-
-      it 'queries the db a reasonable amount' do
-        expect do
-          response, result = post_graphql(**variables) { query }
-          expect(response.status).to eq(200), result.inspect
-          expect(result.dig('data', 'ceOpportunity', 'candidates', 'nodesCount')).to eq(50)
-        end.to make_database_queries(count: 20..30)
-      end
-    end
-
-    context 'when the opportunity has some candidates the current user lacks permission to view' do
-      let(:candidate_pool_with_anonymous) { create :hmis_ce_match_candidate_pool }
-      let(:opportunity) { create :hmis_ce_opportunity, project: project, data_source: ds1, candidate_pool: candidate_pool_with_anonymous }
-
-      let!(:permissioned_project) { create :hmis_hud_project, data_source: ds1, user: u1 }
-      let!(:other_project) { create :hmis_hud_project, data_source: ds1, user: u1 }
-
-      let!(:project_access_control) do
-        create_access_control(
-          hmis_user,
-          permissioned_project,
-          with_permission: [
-            :can_view_clients,
-            :can_view_client_name,
-            :can_view_project,
-            :can_view_enrollment_details,
-          ],
-        )
-      end
-
-      let!(:permissioned_client) { create(:hmis_hud_client_with_warehouse_client, data_source: ds1, first_name: 'Permission Yes', last_name: 'Yep') }
-      let!(:permissioned_enrollment) { create(:hmis_hud_enrollment, client: permissioned_client, project: permissioned_project, data_source: ds1) }
-      let!(:anonymous_client) { create(:hmis_hud_client_with_warehouse_client, data_source: ds1, first_name: 'Permission No', last_name: 'Nope') }
-      let!(:anonymous_enrollment) { create(:hmis_hud_enrollment, client: anonymous_client, project: other_project, data_source: ds1) }
-
-      let!(:permissioned_candidate) do
-        create(
-          :hmis_ce_match_candidate,
-          candidate_pool: candidate_pool_with_anonymous,
-          client: permissioned_client.destination_client,
-          priority_score: 80,
-        )
-      end
-      let!(:anonymous_candidate) do
-        create(
-          :hmis_ce_match_candidate,
-          candidate_pool: candidate_pool_with_anonymous,
-          client: anonymous_client.destination_client,
-          priority_score: 100,
-        )
-      end
-
-      it 'returns some candidates without clients' do
-        response, result = post_graphql(**variables) { query }
-        expect(response.status).to eq(200), result.inspect
-
-        candidates = result.dig('data', 'ceOpportunity', 'candidates', 'nodes')
-        expect(candidates).to contain_exactly(
-          a_hash_including('id' => permissioned_candidate.id.to_s, 'clientName' => 'Permission Yes Yep'),
-          a_hash_including('id' => anonymous_candidate.id.to_s, 'clientName' => "Candidate #{anonymous_candidate.id}"),
-        )
-      end
-    end
-
-    context 'querying for an opportunity that doesnt exist' do
-      let(:variables) do
-        {
-          id: 9999,
-        }
-      end
-
-      it 'does not throw, but returns no opportunity' do
-        response, result = post_graphql(**variables) { query }
-        expect(response.status).to eq(200), result.inspect
-        expect(result.dig('data', 'ceOpportunity')).to be_nil
-      end
-    end
-
-    describe 'without permission' do
-      before do
-        remove_permissions(access_control, :can_view_units)
-      end
-
-      it 'does not return the opportunity' do
-        response, result = post_graphql(**variables) { query }
-        expect(response.status).to eq(200), result.inspect
-        expect(result.dig('data', 'ceOpportunity')).to be_nil
-      end
-    end
-  end
-end
